requirements.txt
python-json-logger
pyyaml
ipykernel